feat(UNT-T16544): code refactor, remove unused assets and update the banner
