+++ .npmignore
@@ -1,4 +1,5 @@
 .github/
+.husky/
 example/
 assets/
 .eslintignore
