+++ example/src/assets/images/index.ts
@@ -1,5 +1,4 @@
 export default {
   forest: require('./forest.jpeg'),
   close: require('./close.png'),
-  flower: require('./flower.jpeg'),
 };
