+++ example/src/components/Types.ts
@@ -1,0 +1,5 @@
+type CustomHeaderType = {
+  title: string;
+};
+
+export type { CustomHeaderType };
