+++ example/src/components/styles/index.ts
@@ -1,0 +1,2 @@
+export { default as CustomHeaderStyle } from './CustomHeaderStyle';
+export { default as ModalHeaderStyle } from './ModalHeaderStyle';
