--- example/src/components/types.ts
@@ -1,5 +1,0 @@
-type CustomHeaderType = {
-  title: string;
-};
-
-export type { CustomHeaderType };
