+++ package.json
@@ -16,11 +16,20 @@
     "react-native",
     "typescript",
     "rn",
-    "images-preview",
-    "images",
-    "preview",
+    "photo",
+    "reanimated",
     "component",
-    "reanimated"
+    "react-component",
+    "iOS",
+    "android",
+    "gallery",
+    "pinch",
+    "pinch-to-zoom",
+    "mobile",
+    "image-zoom",
+    "image",
+    "photo preview",
+    "image preview"
   ],
   "license": "MIT",
   "files": [
