+++ src/components/ImagePreview/Styles.ts
@@ -1,0 +1,21 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  imageParent: {
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  activityIndicatorStyle: {
+    position: 'absolute',
+    top: 0,
+    left: 0,
+    right: 0,
+    bottom: 0,
+  },
+});
+
+export default styles;
