+++ src/components/ImagePreview/index.ts
@@ -1,1 +1,2 @@
 export { default as ImagePreview } from './ImagePreview';
+export * from './Types';
